+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.14.0] - 2019-12-30
 ## Added
 - Add "immediately" parameter to "cancel" method in BillingAgreement
 - Add "cancel_immediately" and "expire" actions to BillingAgreementAdmin
@@ -63,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix missing Refund fields
 
-[Unreleased]: https://github.com/HearthSim/dj-paypal/compare/0.13.0...HEAD
+[Unreleased]: https://github.com/HearthSim/dj-paypal/compare/0.14.0...HEAD
+[0.14.0]: https://github.com/HearthSim/dj-paypal/compare/0.13.0...0.14.0
 [0.13.0]: https://github.com/HearthSim/dj-paypal/compare/0.12.1...0.13.0
 [0.12.1]: https://github.com/HearthSim/dj-paypal/compare/0.12.0...0.12.1
 [0.12.0]: https://github.com/HearthSim/dj-paypal/compare/0.11.2...0.12.0